refactor: move IdleChecker to LiveStreamingServer.Utilities
