+++ src/LiveStreamingServerNet.Rtmp.Relay/Internal/Streams/RtmpUpstreamProcess.cs
@@ -5,8 +5,6 @@ using LiveStreamingServerNet.Rtmp.Relay.Configurations;
 using LiveStreamingServerNet.Rtmp.Relay.Contracts;
 using LiveStreamingServerNet.Rtmp.Relay.Internal.MediaPacketDiscarders.Contracts;
 using LiveStreamingServerNet.Rtmp.Relay.Internal.Streams.Contracts;
-using LiveStreamingServerNet.Rtmp.Relay.Internal.Utilities;
-using LiveStreamingServerNet.Rtmp.Relay.Internal.Utilities.Contracts;
 using LiveStreamingServerNet.Rtmp.Server.Internal.Contracts;
 using LiveStreamingServerNet.Rtmp.Server.Internal.Logging;
 using LiveStreamingServerNet.Rtmp.Server.Internal.Services.Contracts;
