--- src/LiveStreamingServerNet.Rtmp.Relay/Internal/Utilities/Contracts/IIdleChecker.cs
@@ -1,7 +1,0 @@
-﻿namespace LiveStreamingServerNet.Rtmp.Relay.Internal.Utilities.Contracts
-{
-    internal interface IIdleChecker : IDisposable
-    {
-        void Refresh();
-    }
-}
