+++ src/LiveStreamingServerNet.Utilities/Common/Contracts/IIdleChecker.cs
@@ -1,0 +1,7 @@
+﻿namespace LiveStreamingServerNet.Utilities.Common.Contracts
+{
+    public interface IIdleChecker : IDisposable
+    {
+        void Refresh();
+    }
+}
